+++ src/app/component-library/policy-editor/policy-editor.module.ts
@@ -1,17 +1,7 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {MatButtonModule} from '@angular/material/button';
-import {MatChipsModule} from '@angular/material/chips';
-import {MatDatepickerModule} from '@angular/material/datepicker';
-import {MatDialogModule} from '@angular/material/dialog';
-import {MatDividerModule} from '@angular/material/divider';
-import {MatFormFieldModule} from '@angular/material/form-field';
-import {MatIconModule} from '@angular/material/icon';
-import {MatInputModule} from '@angular/material/input';
-import {MatMenuModule} from '@angular/material/menu';
-import {MatSelectModule} from '@angular/material/select';
-import {MatTooltipModule} from '@angular/material/tooltip';
+import {SharedModule} from 'src/app/shared.module';
 import {PipesAndDirectivesModule} from '../pipes-and-directives/pipes-and-directives.module';
 import {ParticipantIdSelectComponent} from './editor/controls/participant-id-select/participant-id-select.component';
 import {PolicyFormAddMenuComponent} from './editor/policy-form-add-menu/policy-form-add-menu.component';
@@ -34,18 +24,7 @@ import {PolicyRendererComponent} from './renderer/policy-renderer/policy-rendere
     ReactiveFormsModule,
 
     // Angular Material
-    MatChipsModule,
-    MatTooltipModule,
-    MatButtonModule,
-    MatDialogModule,
-    MatDividerModule,
-    MatDatepickerModule,
-    MatFormFieldModule,
-    MatMenuModule,
-    MatSelectModule,
-    MatTooltipModule,
-    MatInputModule,
-    MatIconModule,
+    SharedModule,
 
     // EDC UI Modules
     PipesAndDirectivesModule,